+++ package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "@leenguyen/react-flip-clock-countdown": "^1.6.0",
     "babel-runtime": "^6.26.0",
+    "dotenv": "^17.2.3",
     "framer-motion": "^11.15.0",
     "react": "^18.3.1",
     "react-card-flip": "^1.2.3",
